Multiple random object location
Created a prefab to use as the trash. In the Unity inspector, you can adjust the amount of trash, as well as the minimum distance between them spawning.
The script makes attempts to avoid overlap between objects positions.
If failed, it issues a warning in the Unity console.
To do: objects should not overlap with walls.